refactor: decompose App.tsx and extract per-module hooks
