+++ src/app/components/CloseDialogs.tsx
@@ -1,0 +1,121 @@
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "@/components/ui/alert-dialog";
+import type { Tab } from "@/modules/tabs";
+
+type Props = {
+  tabs: Tab[];
+  pendingCloseTab: number | null;
+  onCancelClose: () => void;
+  onConfirmClose: () => void;
+  pendingTerminalCloseTab: number | null;
+  onCancelTerminalClose: () => void;
+  onConfirmTerminalClose: () => void;
+  pendingDeleteTabs: number[] | null;
+  onCancelDeleteClose: () => void;
+  onConfirmDeleteClose: () => void;
+};
+
+/** Confirmation dialogs for closing dirty editors and terminals with live processes. */
+export function CloseDialogs({
+  tabs,
+  pendingCloseTab,
+  onCancelClose,
+  onConfirmClose,
+  pendingTerminalCloseTab,
+  onCancelTerminalClose,
+  onConfirmTerminalClose,
+  pendingDeleteTabs,
+  onCancelDeleteClose,
+  onConfirmDeleteClose,
+}: Props) {
+  return (
+    <>
+      <AlertDialog
+        open={pendingCloseTab !== null}
+        onOpenChange={(open) => !open && onCancelClose()}
+      >
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Unsaved Changes</AlertDialogTitle>
+            <AlertDialogDescription>
+              {tabs.find((t) => t.id === pendingCloseTab)?.title
+                ? `"${
+                    tabs.find((t) => t.id === pendingCloseTab)?.title
+                  }" has unsaved changes. Close anyway?`
+                : "This file has unsaved changes. Close anyway?"}
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel onClick={onCancelClose}>
+              Cancel
+            </AlertDialogCancel>
+            <AlertDialogAction onClick={onConfirmClose}>
+              Close Anyway
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+
+      <AlertDialog
+        open={pendingTerminalCloseTab !== null}
+        onOpenChange={(open) => !open && onCancelTerminalClose()}
+      >
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Close Terminal?</AlertDialogTitle>
+            <AlertDialogDescription>
+              A process is running. Closing this tab will terminate it.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel onClick={onCancelTerminalClose}>
+              Cancel
+            </AlertDialogCancel>
+            <AlertDialogAction onClick={onConfirmTerminalClose}>
+              Close Anyway
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+
+      <AlertDialog
+        open={pendingDeleteTabs !== null}
+        onOpenChange={(open) => !open && onCancelDeleteClose()}
+      >
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Unsaved Changes</AlertDialogTitle>
+            <AlertDialogDescription>
+              {pendingDeleteTabs?.length === 1
+                ? (() => {
+                    const title = tabs.find(
+                      (t) => t.id === pendingDeleteTabs[0],
+                    )?.title;
+                    return title
+                      ? `"${title}" has unsaved changes. The file has been deleted. Close anyway?`
+                      : "This file has unsaved changes. The file has been deleted. Close anyway?";
+                  })()
+                : `${pendingDeleteTabs?.length ?? 0} files have unsaved changes. They have been deleted. Close all anyway?`}
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel onClick={onCancelDeleteClose}>
+              Cancel
+            </AlertDialogCancel>
+            <AlertDialogAction onClick={onConfirmDeleteClose}>
+              Close Anyway
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    </>
+  );
+}
