+++ src/app/components/WorkspaceSurface.tsx
@@ -1,0 +1,165 @@
+import type { ComponentProps } from "react";
+import { cn } from "@/lib/utils";
+import { AiDiffStack, EditorStack, GitDiffStack } from "@/modules/editor";
+import { GitHistoryStack } from "@/modules/git-history";
+import { MarkdownStack } from "@/modules/markdown";
+import { PreviewStack } from "@/modules/preview";
+import type { Tab } from "@/modules/tabs";
+import { TerminalStack } from "@/modules/terminal";
+
+type TerminalStackProps = ComponentProps<typeof TerminalStack>;
+type EditorStackProps = ComponentProps<typeof EditorStack>;
+type PreviewStackProps = ComponentProps<typeof PreviewStack>;
+type AiDiffStackProps = ComponentProps<typeof AiDiffStack>;
+type GitHistoryStackProps = ComponentProps<typeof GitHistoryStack>;
+
+type Props = {
+  tabs: Tab[];
+  activeId: number;
+  activeTab: Tab | undefined;
+  registerTerminalHandle: TerminalStackProps["registerHandle"];
+  onSearchReady: TerminalStackProps["onSearchReady"];
+  onCwd: TerminalStackProps["onCwd"];
+  onExit: TerminalStackProps["onExit"];
+  onFocusLeaf: TerminalStackProps["onFocusLeaf"];
+  registerEditorHandle: EditorStackProps["registerHandle"];
+  onEditorDirtyChange: EditorStackProps["onDirtyChange"];
+  onEditorCloseTab: EditorStackProps["onCloseTab"];
+  registerPreviewHandle: PreviewStackProps["registerHandle"];
+  onPreviewUrlChange: PreviewStackProps["onUrlChange"];
+  onAiDiffAccept: AiDiffStackProps["onAccept"];
+  onAiDiffReject: AiDiffStackProps["onReject"];
+  onOpenCommitFile: GitHistoryStackProps["onOpenCommitFile"];
+  onGitHistorySearchHandle: GitHistoryStackProps["onSearchHandle"];
+};
+
+/**
+ * Stacks every tab-kind surface absolutely on top of each other and toggles
+ * visibility off the active tab, so panes keep their mounted state (terminal
+ * buffers, editor scroll, ...) when switching tabs.
+ */
+export function WorkspaceSurface({
+  tabs,
+  activeId,
+  activeTab,
+  registerTerminalHandle,
+  onSearchReady,
+  onCwd,
+  onExit,
+  onFocusLeaf,
+  registerEditorHandle,
+  onEditorDirtyChange,
+  onEditorCloseTab,
+  registerPreviewHandle,
+  onPreviewUrlChange,
+  onAiDiffAccept,
+  onAiDiffReject,
+  onOpenCommitFile,
+  onGitHistorySearchHandle,
+}: Props) {
+  const kind = activeTab?.kind;
+  const isTerminalTab = kind === "terminal";
+  const isEditorTab = kind === "editor";
+  const isPreviewTab = kind === "preview";
+  const isMarkdownTab = kind === "markdown";
+  const isAiDiffTab = kind === "ai-diff";
+  const isGitDiffTab = kind === "git-diff" || kind === "git-commit-file";
+  const isGitHistoryTab = kind === "git-history";
+
+  return (
+    <div className="relative h-full min-h-0">
+      <div
+        className={cn(
+          "absolute inset-0 px-3 pt-2 pb-2",
+          !isTerminalTab && "invisible pointer-events-none",
+        )}
+        aria-hidden={!isTerminalTab}
+      >
+        <TerminalStack
+          tabs={tabs}
+          activeId={activeId}
+          registerHandle={registerTerminalHandle}
+          onSearchReady={onSearchReady}
+          onCwd={onCwd}
+          onExit={onExit}
+          onFocusLeaf={onFocusLeaf}
+        />
+      </div>
+      <div
+        className={cn(
+          "absolute inset-0 px-3 pt-2 pb-2",
+          !isEditorTab && "invisible pointer-events-none",
+        )}
+        aria-hidden={!isEditorTab}
+      >
+        <EditorStack
+          tabs={tabs}
+          activeId={activeId}
+          registerHandle={registerEditorHandle}
+          onDirtyChange={onEditorDirtyChange}
+          onCloseTab={onEditorCloseTab}
+        />
+      </div>
+      <div
+        className={cn(
+          "absolute inset-0 px-3 pt-2 pb-2",
+          !isPreviewTab && "invisible pointer-events-none",
+        )}
+        aria-hidden={!isPreviewTab}
+      >
+        <PreviewStack
+          tabs={tabs}
+          activeId={activeId}
+          registerHandle={registerPreviewHandle}
+          onUrlChange={onPreviewUrlChange}
+        />
+      </div>
+      <div
+        className={cn(
+          "absolute inset-0 px-3 pt-2 pb-2",
+          !isMarkdownTab && "invisible pointer-events-none",
+        )}
+        aria-hidden={!isMarkdownTab}
+      >
+        <MarkdownStack tabs={tabs} activeId={activeId} />
+      </div>
+      <div
+        className={cn(
+          "absolute inset-0 px-3 pt-2 pb-2",
+          !isAiDiffTab && "invisible pointer-events-none",
+        )}
+        aria-hidden={!isAiDiffTab}
+      >
+        <AiDiffStack
+          tabs={tabs}
+          activeId={activeId}
+          onAccept={onAiDiffAccept}
+          onReject={onAiDiffReject}
+        />
+      </div>
+      <div
+        className={cn(
+          "absolute inset-0 px-3 pt-2 pb-2",
+          !isGitDiffTab && "invisible pointer-events-none",
+        )}
+        aria-hidden={!isGitDiffTab}
+      >
+        <GitDiffStack tabs={tabs} activeId={activeId} />
+      </div>
+      <div
+        className={cn(
+          "absolute inset-0",
+          !isGitHistoryTab && "invisible pointer-events-none",
+        )}
+        aria-hidden={!isGitHistoryTab}
+      >
+        <GitHistoryStack
+          tabs={tabs}
+          activeId={activeId}
+          onOpenCommitFile={onOpenCommitFile}
+          onSearchHandle={onGitHistorySearchHandle}
+        />
+      </div>
+    </div>
+  );
+}
