+++ src/main.tsx
@@ -16,6 +16,13 @@ if (USE_CUSTOM_WINDOW_CONTROLS) {
   document.documentElement.dataset.chrome = "borderless";
 }
 
+// Render-instrumentation overlay, opt-in: `VITE_REACT_SCAN=true pnpm dev`.
+// Dev-only dynamic import so it never reaches the production bundle.
+if (import.meta.env.DEV && import.meta.env.VITE_REACT_SCAN === "true") {
+  const { scan } = await import("react-scan");
+  scan({ enabled: true });
+}
+
 // Reap PTY sessions orphaned by a prior webview load before any tab spawns.
 await invoke("pty_close_all").catch(() => {});
