+++ src/modules/ai/store/chatStore.ts
@@ -1,8 +1,4 @@
-import { Chat, type UIMessage } from "@ai-sdk/react";
-import {
-  type ChatTransport,
-  lastAssistantMessageIsCompleteWithApprovalResponses,
-} from "ai";
+import type { Chat, UIMessage } from "@ai-sdk/react";
 import { create } from "zustand";
 import {
   DEFAULT_MODEL_ID,
@@ -13,10 +9,6 @@ import {
   type ModelId,
   type ProviderId,
 } from "../config";
-import { usePreferencesStore } from "@/modules/settings/preferences";
-import { BUILTIN_AGENTS } from "../lib/agents";
-import { useAgentsStore } from "./agentsStore";
-import { usePlanStore } from "./planStore";
 import { useTodosStore } from "./todoStore";
 import type { AgentUsage } from "../lib/agent";
 import { EMPTY_PROVIDER_KEYS, type ProviderKeys, type CustomEndpointKeys } from "../lib/keyring";
@@ -32,10 +24,8 @@ import {
   type SessionMeta,
 } from "../lib/sessions";
 import { pushRecentModel } from "../lib/modelPrefs";
-import { createContextAwareTransport } from "../lib/transport";
-import type { ToolContext } from "../tools/tools";
 
-type Live = {
+export type Live = {
   getCwd: () => string | null;
   getTerminalContext: () => string | null;
   isActiveTerminalPrivate: () => boolean;
@@ -174,9 +164,9 @@ const NOOP_LIVE: Live = {
 };
 
 const CHATS_LRU_CAP = 8;
-const chats = new Map<string, Chat<UIMessage>>();
+export const chats = new Map<string, Chat<UIMessage>>();
 
-function touchChat(id: string, c: Chat<UIMessage>) {
+export function touchChat(id: string, c: Chat<UIMessage>) {
   if (chats.has(id)) chats.delete(id);
   chats.set(id, c);
   while (chats.size > CHATS_LRU_CAP) {
@@ -190,7 +180,7 @@ function touchChat(id: string, c: Chat<UIMessage>) {
 }
 // Initial messages for a session, populated at hydration time and consumed
 // when the matching Chat is constructed.
-const seedMessages = new Map<string, UIMessage[]>();
+export const seedMessages = new Map<string, UIMessage[]>();
 
 // Trailing debounce for per-token message persistence. Streaming fires
 // `persistMessages` on every token; without this we'd JSON-serialize the
@@ -216,109 +206,6 @@ export function flushPersist(id?: string): void {
     return;
   }
   for (const key of Array.from(pendingPersist.keys())) flushPersistEntry(key);
-}
-
-function makeChat(sessionId: string): Chat<UIMessage> {
-  const readCache = new Map<string, { size: number; hash: number }>();
-  const toolContext: ToolContext = {
-    getCwd: () => useChatStore.getState().live.getCwd(),
-    getWorkspaceRoot: () =>
-      useChatStore.getState().live.getWorkspaceRoot(),
-    getTerminalContext: () =>
-      useChatStore.getState().live.getTerminalContext(),
-    isActiveTerminalPrivate: () =>
-      useChatStore.getState().live.isActiveTerminalPrivate(),
-    injectIntoActivePty: (text) =>
-      useChatStore.getState().live.injectIntoActivePty(text),
-    openPreview: (url) => useChatStore.getState().live.openPreview(url),
-    spawnAgent: (prompt) =>
-      useChatStore.getState().live.spawnManagedAgent(prompt, sessionId),
-    readAgentOutput: (leafId) =>
-      useChatStore.getState().live.readLeafBuffer(leafId),
-    readCache,
-    getSessionId: () => sessionId,
-  };
-
-  const transport = createContextAwareTransport({
-    getKeys: () => useChatStore.getState().apiKeys,
-    toolContext,
-    getModelId: () => useChatStore.getState().selectedModelId,
-    getCustomInstructions: () =>
-      usePreferencesStore.getState().customInstructions,
-    getAgentPersona: () => {
-      const { activeId, customAgents } = useAgentsStore.getState();
-      const all = [...BUILTIN_AGENTS, ...customAgents];
-      const a = all.find((x) => x.id === activeId) ?? BUILTIN_AGENTS[0];
-      return { name: a.name, instructions: a.instructions };
-    },
-    getLive: () => {
-      const live = useChatStore.getState().live;
-      return {
-        cwd: live.getCwd(),
-        terminalPrivate: live.isActiveTerminalPrivate(),
-        workspaceRoot: live.getWorkspaceRoot(),
-        activeFile: live.getActiveFile(),
-      };
-    },
-    getPlanMode: () => usePlanStore.getState().active,
-    getLmstudioBaseURL: () => usePreferencesStore.getState().lmstudioBaseURL,
-    getLmstudioModelId: () => usePreferencesStore.getState().lmstudioModelId,
-    getMlxBaseURL: () => usePreferencesStore.getState().mlxBaseURL,
-    getMlxModelId: () => usePreferencesStore.getState().mlxModelId,
-    getOllamaBaseURL: () => usePreferencesStore.getState().ollamaBaseURL,
-    getOllamaModelId: () => usePreferencesStore.getState().ollamaModelId,
-    getOpenaiCompatibleBaseURL: () =>
-      usePreferencesStore.getState().openaiCompatibleBaseURL,
-    getOpenaiCompatibleModelId: () =>
-      usePreferencesStore.getState().openaiCompatibleModelId,
-    getOpenaiCompatibleContextLimit: () =>
-      usePreferencesStore.getState().openaiCompatibleContextLimit,
-    getOpenrouterModelId: () =>
-      usePreferencesStore.getState().openrouterModelId,
-    getCustomEndpoints: () =>
-      usePreferencesStore.getState().customEndpoints,
-    getCustomEndpointKeys: () =>
-      useChatStore.getState().customEndpointKeys,
-    onStep: (step) => {
-      useChatStore.getState().patchAgentMeta({ step });
-    },
-    onCompact: (info) => {
-      useChatStore.getState().patchAgentMeta({
-        compactionNotice: { droppedCount: info.droppedCount, at: Date.now() },
-      });
-    },
-    onFinishMeta: (info) => {
-      useChatStore.getState().patchAgentMeta({ hitStepCap: info.hitStepCap });
-    },
-    onUsage: (delta) => {
-      const cur = useChatStore.getState().agentMeta.tokens;
-      useChatStore.getState().patchAgentMeta({
-        tokens: {
-          inputTokens: cur.inputTokens + delta.inputTokens,
-          outputTokens: cur.outputTokens + delta.outputTokens,
-          cachedInputTokens: cur.cachedInputTokens + delta.cachedInputTokens,
-        },
-        lastInputTokens: delta.lastInputTokens,
-        lastCachedTokens: delta.lastCachedTokens,
-      });
-    },
-  }) as unknown as ChatTransport<UIMessage>;
-
-  const initialMessages = seedMessages.get(sessionId);
-  seedMessages.delete(sessionId);
-
-  return new Chat<UIMessage>({
-    id: sessionId,
-    transport,
-    messages: initialMessages,
-    sendAutomaticallyWhen: lastAssistantMessageIsCompleteWithApprovalResponses,
-    onError: (e) => {
-      useChatStore.getState().patchAgentMeta({
-        status: "error",
-        error: e instanceof Error ? e.message : String(e),
-      });
-    },
-  });
 }
 
 export const useChatStore = create<StoreState>((set, get) => ({
@@ -558,31 +445,10 @@ export function hasKeyForModel(modelId: string): boolean {
   return providerNeedsKey(provider) ? !!apiKeys[provider] : true;
 }
 
-export function getOrCreateChat(sessionId: string): Chat<UIMessage> {
-  const existing = chats.get(sessionId);
-  if (existing) {
-    touchChat(sessionId, existing);
-    return existing;
-  }
-  const c = makeChat(sessionId);
-  touchChat(sessionId, c);
-  return c;
-}
-
 export function getChat(sessionId?: string): Chat<UIMessage> | undefined {
   if (sessionId) return chats.get(sessionId);
   const id = useChatStore.getState().activeSessionId;
   return id ? chats.get(id) : undefined;
-}
-
-export async function sendMessage(text: string): Promise<boolean> {
-  const state = useChatStore.getState();
-  const sessionId = state.activeSessionId;
-  if (!sessionId) return false;
-  if (providerNeedsKey(getModel(state.selectedModelId as ModelId).provider) && !getActiveProviderKey()) return false;
-  const c = getOrCreateChat(sessionId);
-  await c.sendMessage({ text });
-  return true;
 }
 
 export function stop(): void {
