+++ src/modules/editor/index.ts
@@ -3,3 +3,4 @@ export { EditorStack } from "./EditorStackLazy";
 export { AiDiffStack } from "./AiDiffStackLazy";
 export { GitDiffStack } from "./GitDiffStackLazy";
 export { NewEditorDialog } from "./NewEditorDialog";
+export { useEditorFileSync } from "./useEditorFileSync";
