+++ src/modules/editor/lib/colorSwatches.ts
@@ -2,7 +2,7 @@ import { RangeSetBuilder, type Extension } from "@codemirror/state";
 import {
   Decoration,
   type DecorationSet,
-  EditorView,
+  type EditorView,
   ViewPlugin,
   type ViewUpdate,
   WidgetType,
