+++ src/modules/preview/index.ts
@@ -1,2 +1,2 @@
 export { PreviewStack } from "./PreviewStack";
-export { type PreviewPaneHandle } from "./PreviewPane";
+export type { PreviewPaneHandle } from "./PreviewPane";
