+++ src/modules/sidebar/index.ts
@@ -1,2 +1,7 @@
 export { SidebarRail, SIDEBAR_RAIL_HEIGHT } from "./SidebarRail";
 export type { SidebarViewId } from "./types";
+export {
+  useSidebarPanel,
+  SIDEBAR_MIN_WIDTH,
+  SIDEBAR_MAX_WIDTH,
+} from "./useSidebarPanel";
