+++ src/modules/sidebar/useSidebarPanel.ts
@@ -1,0 +1,160 @@
+import {
+  type RefObject,
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
+import type { PanelImperativeHandle } from "react-resizable-panels";
+import type { SidebarViewId } from "./types";
+
+export const SIDEBAR_DEFAULT_WIDTH = 260;
+export const SIDEBAR_MIN_WIDTH = 220;
+export const SIDEBAR_MAX_WIDTH = 480;
+const SIDEBAR_WIDTH_STORAGE_KEY = "terax.sidebar.width";
+const SIDEBAR_VIEW_STORAGE_KEY = "terax.sidebar.view";
+
+function clampSidebarWidth(width: number): number {
+  return Math.min(
+    SIDEBAR_MAX_WIDTH,
+    Math.max(SIDEBAR_MIN_WIDTH, Math.round(width)),
+  );
+}
+
+function readSidebarWidth(): number {
+  try {
+    const stored = window.localStorage.getItem(SIDEBAR_WIDTH_STORAGE_KEY);
+    const parsed = stored ? Number.parseInt(stored, 10) : NaN;
+    return Number.isFinite(parsed)
+      ? clampSidebarWidth(parsed)
+      : SIDEBAR_DEFAULT_WIDTH;
+  } catch {
+    return SIDEBAR_DEFAULT_WIDTH;
+  }
+}
+
+function readSidebarView(): SidebarViewId {
+  try {
+    const stored = window.localStorage.getItem(SIDEBAR_VIEW_STORAGE_KEY);
+    if (stored === "explorer" || stored === "source-control") return stored;
+  } catch {
+    // ignore
+  }
+  return "explorer";
+}
+
+type FocusableExplorer = {
+  focus: () => void;
+  isFocused: () => boolean;
+};
+
+export function useSidebarPanel(
+  explorerRef: RefObject<FocusableExplorer | null>,
+) {
+  const sidebarRef = useRef<PanelImperativeHandle | null>(null);
+  const sidebarWidthRef = useRef(readSidebarWidth());
+  const sidebarWidthWriteTimerRef = useRef(0);
+  const explorerReturnFocusRef = useRef<HTMLElement | null>(null);
+  const [sidebarView, setSidebarViewState] =
+    useState<SidebarViewId>(readSidebarView);
+
+  const persistSidebarView = useCallback((view: SidebarViewId) => {
+    setSidebarViewState(view);
+    try {
+      window.localStorage.setItem(SIDEBAR_VIEW_STORAGE_KEY, view);
+    } catch {
+      // storage may fail in private mode
+    }
+  }, []);
+
+  const toggleSidebar = useCallback(() => {
+    const p = sidebarRef.current;
+    if (!p) return;
+    if (p.getSize().asPercentage <= 0) p.expand();
+    else p.collapse();
+  }, []);
+
+  const cycleSidebarView = useCallback(
+    (view: SidebarViewId) => {
+      const panel = sidebarRef.current;
+      const collapsed = panel ? panel.getSize().asPercentage <= 0 : false;
+      if (collapsed) {
+        if (panel) panel.resize(`${sidebarWidthRef.current}px`);
+        if (view !== sidebarView) persistSidebarView(view);
+        return;
+      }
+      if (view === sidebarView) {
+        panel?.collapse();
+        return;
+      }
+      persistSidebarView(view);
+    },
+    [persistSidebarView, sidebarView],
+  );
+
+  const persistSidebarWidth = useCallback((next: number) => {
+    sidebarWidthRef.current = next;
+    if (sidebarWidthWriteTimerRef.current) {
+      window.clearTimeout(sidebarWidthWriteTimerRef.current);
+    }
+    sidebarWidthWriteTimerRef.current = window.setTimeout(() => {
+      sidebarWidthWriteTimerRef.current = 0;
+      try {
+        window.localStorage.setItem(SIDEBAR_WIDTH_STORAGE_KEY, String(next));
+      } catch {
+        // ignore
+      }
+    }, 200);
+  }, []);
+
+  useEffect(() => {
+    return () => {
+      if (sidebarWidthWriteTimerRef.current) {
+        window.clearTimeout(sidebarWidthWriteTimerRef.current);
+      }
+    };
+  }, []);
+
+  const toggleExplorerFocus = useCallback(() => {
+    const explorer = explorerRef.current;
+    const panel = sidebarRef.current;
+    const collapsed = panel ? panel.getSize().asPercentage <= 0 : false;
+    if (sidebarView !== "explorer" || collapsed) {
+      if (panel && collapsed) panel.resize(`${sidebarWidthRef.current}px`);
+      if (sidebarView !== "explorer") persistSidebarView("explorer");
+      const active = document.activeElement;
+      explorerReturnFocusRef.current =
+        active instanceof HTMLElement && active !== document.body
+          ? active
+          : null;
+      requestAnimationFrame(() => explorerRef.current?.focus());
+      return;
+    }
+    if (!explorer) return;
+    if (explorer.isFocused()) {
+      const target = explorerReturnFocusRef.current;
+      explorerReturnFocusRef.current = null;
+      if (target && document.body.contains(target)) {
+        target.focus();
+      } else {
+        (document.activeElement as HTMLElement | null)?.blur?.();
+      }
+      return;
+    }
+    const active = document.activeElement;
+    explorerReturnFocusRef.current =
+      active instanceof HTMLElement && active !== document.body ? active : null;
+    explorer.focus();
+  }, [explorerRef, persistSidebarView, sidebarView]);
+
+  return {
+    sidebarRef,
+    sidebarWidthRef,
+    sidebarView,
+    persistSidebarView,
+    toggleSidebar,
+    cycleSidebarView,
+    persistSidebarWidth,
+    toggleExplorerFocus,
+  };
+}
