+++ src/modules/source-control/index.ts
@@ -4,3 +4,4 @@ export {
   useSourceControl,
   type SourceControlSummary,
 } from "./useSourceControl";
+export { useSourceControlContext } from "./useSourceControlContext";
