+++ src/modules/source-control/useSourceControlContext.ts
@@ -1,0 +1,111 @@
+import { useCallback, useMemo } from "react";
+import { native } from "@/modules/ai/lib/native";
+import type { SidebarViewId } from "@/modules/sidebar";
+import type { Tab } from "@/modules/tabs";
+import { useSourceControl } from "./useSourceControl";
+
+function dirname(path: string | null): string | null {
+  if (!path) return null;
+  const normalized = path.replace(/\\/g, "/");
+  const idx = normalized.lastIndexOf("/");
+  if (idx <= 0) return normalized;
+  return normalized.slice(0, idx);
+}
+
+type Params = {
+  activeTab: Tab | undefined;
+  tabs: Tab[];
+  activeTerminalLeafCwd: string | null;
+  explorerRoot: string | null;
+  launchCwd: string | null;
+  launchCwdResolved: boolean;
+  home: string | null;
+  sidebarView: SidebarViewId;
+  cycleSidebarView: (view: SidebarViewId) => void;
+  openCommitHistoryTab: (args: {
+    repoRoot: string;
+    branch: string | null;
+  }) => void;
+};
+
+/**
+ * Resolves the source-control context path off the active tab and feeds the
+ * source-control summary. When git is not active the badge tracks a stable
+ * per-session path so tab switches / cd don't re-fire git IPC.
+ */
+export function useSourceControlContext({
+  activeTab,
+  tabs,
+  activeTerminalLeafCwd,
+  explorerRoot,
+  launchCwd,
+  launchCwdResolved,
+  home,
+  sidebarView,
+  cycleSidebarView,
+  openCommitHistoryTab,
+}: Params) {
+  const workspaceFallbackPath = launchCwdResolved
+    ? (launchCwd ?? home ?? null)
+    : null;
+  const sourceControlContextPath = (() => {
+    if (activeTab?.kind === "terminal") {
+      return activeTerminalLeafCwd ?? explorerRoot ?? workspaceFallbackPath;
+    }
+    if (activeTab?.kind === "editor") return dirname(activeTab.path);
+    if (activeTab?.kind === "git-diff") return activeTab.repoRoot;
+    if (activeTab?.kind === "git-commit-file") return activeTab.repoRoot;
+    if (activeTab?.kind === "git-history") return activeTab.repoRoot;
+    return explorerRoot ?? workspaceFallbackPath;
+  })();
+  const hasOpenGitTab = useMemo(
+    () =>
+      tabs.some(
+        (t) =>
+          t.kind === "git-diff" ||
+          t.kind === "git-history" ||
+          t.kind === "git-commit-file",
+      ),
+    [tabs],
+  );
+  const sourceControlActive = hasOpenGitTab || sidebarView === "source-control";
+  // Stable per-session path so switching tabs / cd-ing in a shell does NOT
+  // re-fire git IPC for the badge. The active panel resolves the current
+  // context path on its own when the user actually opens git.
+  const badgeContextPath = workspaceFallbackPath;
+  const sourceControlPath = sourceControlActive
+    ? sourceControlContextPath
+    : badgeContextPath;
+  const sourceControl = useSourceControl(sourceControlPath, true);
+
+  const toggleSourceControl = useCallback(() => {
+    cycleSidebarView("source-control");
+  }, [cycleSidebarView]);
+
+  const openGitGraphFromContext = useCallback(async () => {
+    const known = sourceControl.hasRepo ? sourceControl.repo : null;
+    if (known) {
+      openCommitHistoryTab({
+        repoRoot: known.repoRoot,
+        branch: sourceControl.status?.branch ?? null,
+      });
+      return;
+    }
+    if (!sourceControlContextPath) return;
+    try {
+      const repo = await native.gitResolveRepo(sourceControlContextPath);
+      if (!repo) return;
+      openCommitHistoryTab({ repoRoot: repo.repoRoot, branch: repo.branch });
+    } catch {
+      /* noop */
+    }
+  }, [
+    openCommitHistoryTab,
+    sourceControl.hasRepo,
+    sourceControl.repo,
+    sourceControl.status?.branch,
+    sourceControlContextPath,
+  ]);
+
+  return { sourceControl, toggleSourceControl, openGitGraphFromContext };
+}
