+++ src/modules/theme/index.ts
@@ -1,1 +1,2 @@
 export { ThemeProvider, useTheme, type Theme } from "./ThemeProvider";
+export { useThemeFileEditing } from "./useThemeFileEditing";
